Move check to cppcoreguidelines- section
