+++ clang-tools-extra/clang-tidy/modernize/ModernizeTidyModule.cpp
@@ -35,7 +35,6 @@
 #include "UseDefaultMemberInitCheck.h"
 #include "UseDesignatedInitializersCheck.h"
 #include "UseEmplaceCheck.h"
-#include "UseEnumClassCheck.h"
 #include "UseEqualsDefaultCheck.h"
 #include "UseEqualsDeleteCheck.h"
 #include "UseIntegerSignComparisonCheck.h"
@@ -112,8 +111,7 @@ public:
         "modernize-use-default-member-init");
     CheckFactories.registerCheck<UseEmplaceCheck>("modernize-use-emplace");
     CheckFactories.registerCheck<UseEnumClassCheck>("modernize-use-enum-class");
-    CheckFactories.registerCheck<UseEqualsDefaultCheck>(
-        "modernize-use-equals-default");
+    CheckFactories.registerCheck<UseEqualsDefaultCheck>("modernize-use-equals-default");
     CheckFactories.registerCheck<UseEqualsDeleteCheck>(
         "modernize-use-equals-delete");
     CheckFactories.registerCheck<UseNodiscardCheck>("modernize-use-nodiscard");
